set up parallax scrolling #175
