+++ package.json
@@ -40,6 +40,7 @@
     "clipboard": "^1.5.12",
     "object-fit-images": "^3.2.4",
     "picturefill": "*",
+    "rellax": "^1.12.1",
     "smooth-scroll": "^16.1.3"
   },
   "devDependencies": {
