+++ source/assets/stylesheets/components/_button.scss
@@ -8,6 +8,7 @@
 @import '../config/scale';
 @import '../config/type';
 @import '../config/borders';
+@import '../config/special_effects';
 
 $button-icon-size: 1em;
 $button-icon-margin: ms(0, 1em);
