+++ source/assets/stylesheets/utilities/_positioning.scss
@@ -59,6 +59,10 @@
 	overflow: hidden;
 }
 
+.overflow-visible {
+	overflow: visible;
+}
+
 .align-center-block {
 	margin-left: auto;
 	margin-right: auto;
